Logo path logic.
Added logo management logic to the project. The user is now able to add a logo when creating a new company entry and is able to edit an existing companies logo in the edit view. Logo's are stored in the correct file and are displayed correctly on the index and show views. The edit view displays the current companies logo under the add logo form. Updated the company seeder with a default image for each default company.