show view rework.
Reworked the layout of the company and employee show views. The show views now use a card layout that show's all the information of the selected company or employee as well as having all action buttons on the bottom right side of the card layout. The employee card now displays the name of the company it is assigned to. added the employee index view card layout to the company show view's employee listing. The company show views card layout displays the selected company logo on the left side of the card layout. Updated the employee controllers logic to display the employee companies name on the show view. updated the site CSS file with the new CSS for updated card layout.